made App.js stateless -- all components now managing their own state
